move memory class out PVZ.h
