+++ pvzclass/pvzclass.vcxproj.filters
@@ -737,5 +737,8 @@
     <ClInclude Include="Classes\SeedChooserScreen.hpp">
       <Filter>头文件\Classes</Filter>
     </ClInclude>
+    <ClInclude Include="Memory.hpp">
+      <Filter>头文件</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
